// src/data/terminals.generated.ts
export const TERMINALS: Terminal[] = [
  { id: 'osm-node-5648522790', osmId: 5648522790, name: "Buses Delsal", lat: -36.8218602, lng: -73.0510712 },
  { id: 'osm-node-5648520529', osmId: 5648520529, name: "Conquistador", lat: -36.8217989, lng: -73.0511223 },
  { id: 'osm-node-1611101296', osmId: 1611101296, name: "Estación de buses Turbus", lat: -36.914896, lng: -73.028563 },
  { id: 'osm-way-135488014', osmId: 135488014, name: "Estación Intermodal Concepción", lat: -36.8298968, lng: -73.0605492 },
  { id: 'osm-node-6718688807', osmId: 6718688807, name: "Estación Intermodal Coronel", lat: -37.0246243, lng: -73.1505317 },
  { id: 'osm-way-726995915', osmId: 726995915, name: "Terminal Buses Geminis Sur", lat: -36.8215704, lng: -72.9930842 },
  { id: 'osm-node-1249412543', osmId: 1249412543, name: "Terminal de Buses 30 Ruta las Playas", lat: -36.7132656, lng: -72.9773881 },
  { id: 'osm-node-1249412318', osmId: 1249412318, name: "Terminal de Buses 32 Buses Ruta del Mar", lat: -36.710914, lng: -72.9742832 },
  { id: 'osm-node-1409373698', osmId: 1409373698, name: "Terminal de Buses 63 Rengo Lientur", lat: -36.805472, lng: -73.0383875 },
  { id: 'osm-way-114474600', osmId: 114474600, name: "Terminal de Buses Camilo Henríquez", lat: -36.8087472, lng: -73.0340681 },
  { id: 'osm-way-597586612', osmId: 597586612, name: "Terminal de Buses Collao", lat: -36.8158265, lng: -73.0223785 },
  { id: 'osm-way-464601931', osmId: 464601931, name: "Terminal de buses Expresos Chiguayante", lat: -36.9514227, lng: -73.0158939, network: "TPMC", operator: "Expresos Chiguayante" },
  { id: 'osm-way-423901976', osmId: 423901976, name: "Terminal de Buses Jota Ewert", lat: -36.8184045, lng: -73.0601454 },
  { id: 'osm-way-220495893', osmId: 220495893, name: "Terminal de buses Las Bahías", lat: -36.8428323, lng: -73.0099705, network: "TPMC", operator: "Las Bahías" },
  { id: 'osm-way-464462966', osmId: 464462966, name: "Terminal de buses Las Galaxias", lat: -36.9734629, lng: -72.9267408, network: "TPMC", operator: "Las Galaxias" },
  { id: 'osm-way-465064981', osmId: 465064981, name: "Terminal de buses Mi Expreso", lat: -36.7188788, lng: -73.1210552, network: "TPMC", operator: "Mi Expreso" },
  { id: 'osm-node-5966229600', osmId: 5966229600, name: "Terminal de Buses Nueva Llacolén", lat: -36.8288511, lng: -73.1476142 },
  { id: 'osm-node-5355452225', osmId: 5355452225, name: "Terminal de Buses Palomares", lat: -36.9494134, lng: -72.9317719 },
  { id: 'osm-way-464609623', osmId: 464609623, name: "Terminal de buses Pedro de Valdivia", lat: -36.8656908, lng: -73.0330818, network: "TPMC", operator: "Pedro de Valdivia" },
  { id: 'osm-way-465063020', osmId: 465063020, name: "Terminal de buses Puchacay", lat: -36.83921, lng: -73.0082079, network: "TPMC", operator: "Buses Puchacay" },
  { id: 'osm-way-253665135', osmId: 253665135, name: "Terminal de buses Vía del Sol", lat: -36.965035, lng: -72.9227312, network: "TPMC", operator: "Vía del Sol" },
  { id: 'osm-way-464460912', osmId: 464460912, name: "Terminal de buses Vía Futuro", lat: -36.9538151, lng: -73.0221811, network: "TPMC", operator: "Vía Futuro" },
  { id: 'osm-way-464462253', osmId: 464462253, name: "Terminal de buses Vía Universo", lat: -36.9590421, lng: -73.0106787, network: "TPMC", operator: "Vía Universo" },
  { id: 'osm-node-4670412642', osmId: 4670412642, name: "Terminal de colectivos 21", lat: -36.7954695, lng: -73.1071713, operator: "Las Golondrinas" },
  { id: 'osm-way-425356582', osmId: 425356582, name: "Terminal de líneas a Tomé", lat: -36.8191314, lng: -73.0619879 },
  { id: 'osm-way-196575222', osmId: 196575222, name: "Terminal Félix Adán", lat: -36.7111779, lng: -73.1154408 },
  { id: 'osm-node-1430983050', osmId: 1430983050, name: "Terminal Los Lobos", lat: -36.7229162, lng: -73.1430985 },
  { id: 'osm-way-1462655442', osmId: 1462655442, name: "Terminal San Remo", lat: -36.8242286, lng: -73.1301212 },
];